automated data retrieval
I found a table where GiveWell lists the donations they granted individual charites (https://airtable.com/appGuFtOIb1eodoBu/shr1EzngorAlEzziP/tblG72bMUu36lrWsr/viwJScgH8B6uRwJ8C?blocks=hide). This table is copied into "grant_data.xlsx". The GiveWell table Emma suggested is in "GiveWell_table.xlsx". I am now trying to automate the retrieval of the data from the grant_data table to buil a base for our databank, My goal is to save the databank in a separate file so scripts that we develop later can access it quickly and easily.

This is the same commit as the 2nd to last. I want the data stuff in a separate folder and had to figure out how to do it.